Refactor, categorize and pull out function definitions from symbol broker
