+++ NINA.Sequencer/Logic/SymbolBroker.cs
@@ -33,6 +33,7 @@ using NINA.Equipment.Interfaces.Mediator;
 using NINA.Image.ImageData;
 using NINA.Profile.Interfaces;
 using NINA.Sequencer.Conditions;
+using NINA.Sequencer.Logic.SymbolFunctions;
 using NINA.WPF.Base.ViewModel;
 using Parlot.Fluent;
 using System;
@@ -89,8 +90,8 @@ namespace NINA.Sequencer.Logic {
             foreach (string provider in SymbolProviders) {
                 RegisterSymbolProvider(provider);
             }
-            // Register basic functions
-            RegisterBasicFunctions();
+            // Register the core functions
+            RegisterCoreFunctions();
         }
 
         private ConcurrentDictionary<string, IList<Symbol>> DataSymbols = new ConcurrentDictionary<string, IList<Symbol>>();
@@ -633,795 +634,26 @@ namespace NINA.Sequencer.Logic {
             }
         }
 
-        private static Random rng = new Random();
-
-        private void RegisterBasicFunctions() {
-            static DateTime GetDateTime(FunctionArgs args) {
-                DateTime dt;
-                if (args.Parameters?.Length > 0) {
-                    try {
-                        var utc = CoreUtil.UnixTimeStampToDateTime(
-                            Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture));
-                        dt = utc.ToLocalTime();
-                    } catch {
-                        dt = DateTime.MinValue;
-                    }
-                } else {
-                    dt = DateTime.Now;
-                }
-                return dt;
+        private void RegisterCoreFunctions() {
+            foreach (var fn in new MathFunctions(this)) {
+                RegisterFunction(fn.Category, fn);
             }
 
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Abs",
-                    category: "NINA",
-                    description: "Returns the absolute value of a specified number.",
-                    usageExample: "Abs(-1)",
-                    implementation: args => Math.Abs(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
+            foreach (var fn in new LogicFunctions(this)) {
+                RegisterFunction(fn.Category, fn);
+            }
 
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Acos",
-                    category: "NINA",
-                    description: "Returns the angle whose cosine is the specified number.",
-                    usageExample: "Acos(1)",
-                    implementation: args => Math.Acos(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
+            foreach (var fn in new TimeFunctions(this)) {
+                RegisterFunction(fn.Category, fn);
+            }
 
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Asin",
-                    category: "NINA",
-                    description: "Returns the angle whose sine is the specified number.",
-                    usageExample: "Asin(0)",
-                    implementation: args => Math.Asin(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
+            foreach (var fn in new StringFunctions(this)) {
+                RegisterFunction(fn.Category, fn);
+            }
 
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Atan",
-                    category: "NINA",
-                    description: "Returns the angle whose tangent is the specified number.",
-                    usageExample: "Atan(0)",
-                    implementation: args => Math.Atan(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Ceiling",
-                    category: "NINA",
-                    description: "Returns the smallest integer greater than or equal to the specified number.",
-                    usageExample: "Ceiling(1.5)",
-                    implementation: args => Math.Ceiling(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Cos",
-                    category: "NINA",
-                    description: "Returns the cosine of the specified angle.",
-                    usageExample: "Cos(0)",
-                    implementation: args => Math.Cos(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Exp",
-                    category: "NINA",
-                    description: "Returns e raised to the specified power.",
-                    usageExample: "Exp(0)",
-                    implementation: args => Math.Exp(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Floor",
-                    category: "NINA",
-                    description: "Returns the largest integer less than or equal to the specified number.",
-                    usageExample: "Floor(1.5)",
-                    implementation: args => Math.Floor(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "IEEERemainder",
-                    category: "NINA",
-                    description: "Returns the remainder resulting from the division of a specified number by another specified number.",
-                    usageExample: "IEEERemainder(3, 2)",
-                    implementation: args => Math.IEEERemainder(
-                        Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture),
-                        Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 2, maxArgs: 2)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Ln",
-                    category: "NINA",
-                    description: "Returns the natural logarithm of a specified number.",
-                    usageExample: "Ln(1)",
-                    implementation: args => Math.Log(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Log",
-                    category: "NINA",
-                    description: "Returns the logarithm of a specified number.",
-                    usageExample: "Log(1, 10)",
-                    implementation: args => Math.Log(
-                        Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture),
-                        Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 2, maxArgs: 2)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Log10",
-                    category: "NINA",
-                    description: "Returns the base 10 logarithm of a specified number.",
-                    usageExample: "Log10(1)",
-                    implementation: args => Math.Log10(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Max",
-                    category: "NINA",
-                    description: "Returns the larger of two specified numbers.",
-                    usageExample: "Max(1, 2)",
-                    implementation: args => Math.Max(
-                        Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture),
-                        Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 2, maxArgs: 2)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Min",
-                    category: "NINA",
-                    description: "Returns the smaller of two numbers.",
-                    usageExample: "Min(1, 2)",
-                    implementation: args => Math.Min(
-                        Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture),
-                        Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 2, maxArgs: 2)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Pow",
-                    category: "NINA",
-                    description: "Returns a specified number raised to the specified power.",
-                    usageExample: "Pow(3, 2)",
-                    implementation: args => Math.Pow(
-                        Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture),
-                        Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 2, maxArgs: 2)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Round",
-                    category: "NINA",
-                    description: "Rounds a value to the nearest integer or specified number of decimal places.",
-                    usageExample: "Round(3.222, 2)",
-                    implementation: args => {
-                        // 1 or 2 args: Round(x) or Round(x, decimals)
-                        if (args.Parameters.Length == 2) {
-                            return Math.Round(
-                                Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture),
-                                Convert.ToInt32(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture));
-                        }
-
-                        return Math.Round(Convert.ToDouble(args.Parameters[0].Evaluate()));
-                    },
-                    minArgs: 1, maxArgs: 2)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Sign",
-                    category: "NINA",
-                    description: "Returns a value indicating the sign of a number.",
-                    usageExample: "Sign(-10)",
-                    implementation: args => Math.Sign(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Sin",
-                    category: "NINA",
-                    description: "Returns the sine of the specified angle.",
-                    usageExample: "Sin(0)",
-                    implementation: args => Math.Sin(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Sqrt",
-                    category: "NINA",
-                    description: "Returns the square root of a specified number.",
-                    usageExample: "Sqrt(4)",
-                    implementation: args => Math.Sqrt(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Tan",
-                    category: "NINA",
-                    description: "Returns the tangent of the specified angle.",
-                    usageExample: "Tan(0)",
-                    implementation: args => Math.Tan(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Truncate",
-                    category: "NINA",
-                    description: "Calculates the integral part of a number.",
-                    usageExample: "Truncate(1.7)",
-                    implementation: args => Math.Truncate(Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture)),
-                    minArgs: 1, maxArgs: 1)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "In",
-                    category: "NINA",
-                    description: "Returns whether an element is in a set of values.",
-                    usageExample: "in(1 + 1, 1, 2, 3)",
-                    implementation: args => {
-                        var value = args.Parameters[0].Evaluate();
-                        for (int i = 1; i < args.Parameters.Length; i++) {
-                            if (Equals(value, args.Parameters[i].Evaluate()))
-                                return true;
-                        }
-                        return false;
-                    },
-                    minArgs: 2, maxArgs: int.MaxValue)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "If",
-                    category: "NINA",
-                    description: "Returns a value based on a condition.",
-                    usageExample: "if(3 % 2 = 1, 'value is true', 'value is false')",
-                    implementation: args => {
-                        bool condition = Convert.ToBoolean(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        return condition
-                            ? args.Parameters[1].Evaluate()
-                            : args.Parameters[2].Evaluate();
-                    },
-                    minArgs: 3, maxArgs: 3)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Ifs",
-                    category: "NINA",
-                    description: "Returns a value based on evaluating a number of conditions, with a default if none are true.",
-                    usageExample: "ifs(foo > 50, \"bar\", foo > 75, \"baz\", \"quux\")",
-                    implementation: args => {
-                        int count = args.Parameters.Length;
-
-                        // at least condition, value, default
-                        if (count < 3)
-                            throw new ArgumentException("ifs() requires at least 3 arguments.");
-
-                        // all but last are (condition, value) pairs
-                        for (int i = 0; i < count - 1; i += 2) {
-                            bool cond = Convert.ToBoolean(args.Parameters[i].Evaluate(), CultureInfo.InvariantCulture);
-                            if (cond)
-                                return args.Parameters[i + 1].Evaluate();
-                        }
-
-                        // default value (last argument)
-                        return args.Parameters[count - 1].Evaluate();
-                    },
-                    minArgs: 3, maxArgs: int.MaxValue)
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Now",
-                category: "NINA",
-                description: "Returns the current Unix timestamp in seconds.",
-                usageExample: "Now()",
-                implementation: args => CoreUtil.UnixTimeStampNow(),
-                minArgs: 0,
-                maxArgs: 0,
-                isVolatile: true
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Hour",
-                category: "NINA",
-                description: "Returns the hour component (0–23) of a given datetime, or of the current time if no argument is supplied.",
-                usageExample: "Hour() or Hour(someDate)",
-                implementation: args => (int)GetDateTime(args).Hour,
-                minArgs: 0,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Minute",
-                category: "NINA",
-                description: "Returns the minute component (0–59) of a given datetime, or of the current time if no argument is supplied.",
-                usageExample: "Minute() or Minute(someDate)",
-                implementation: args => (int)GetDateTime(args).Minute,
-                minArgs: 0,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Day",
-                category: "NINA",
-                description: "Returns the day of the month (1–31) of a given datetime, or of the current date if no argument is supplied.",
-                usageExample: "Day() or Day(someDate)",
-                implementation: args => (int)GetDateTime(args).Day,
-                minArgs: 0,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Month",
-                category: "NINA",
-                description: "Returns the month (1–12) of a given datetime, or of the current date if no argument is supplied.",
-                usageExample: "Month() or Month(someDate)",
-                implementation: args => (int)GetDateTime(args).Month,
-                minArgs: 0,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Year",
-                category: "NINA",
-                description: "Returns the year component of a given datetime, or of the current date if no argument is supplied.",
-                usageExample: "Year() or Year(someDate)",
-                implementation: args => (int)GetDateTime(args).Year,
-                minArgs: 0,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Dow",
-                category: "NINA",
-                description: "Returns the day of the week as an integer (0 = Sunday, 1 = Monday, … 6 = Saturday).",
-                usageExample: "Dow() or Dow(someDate)",
-                implementation: args => (int)GetDateTime(args).DayOfWeek,
-                minArgs: 0,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "DateString",
-                category: "NINA",
-                description: "Formats a datetime value using the specified .NET format string.",
-                usageExample: "DateString(now(), \"yyyy-MM-dd HH:mm:ss\")",
-                implementation: args => {
-                    var dt = GetDateTime(args);
-                    var fmt = Convert.ToString(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                    return dt.ToString(fmt);
-                },
-                minArgs: 2,
-                maxArgs: 2
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "AddMinutes",
-                    category: "NINA",
-                    description: "Adds minutes to a datetime. Returns a Unix timestamp in seconds.",
-                    usageExample: "AddMinutes(now(), 30)",
-                    implementation: args => {
-                        DateTime baseDt;
-                        double minutes;
-
-                        if (args.Parameters.Length == 1) {
-                            // base: now, arg: minutes
-                            baseDt = DateTime.Now;
-                            minutes = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        } else {
-                            // first arg: unix timestamp, second: minutes
-                            var baseSeconds = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                            baseDt = CoreUtil.UnixTimeStampToDateTime(baseSeconds).ToLocalTime();
-                            minutes = Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        }
-
-                        var result = baseDt.AddMinutes(minutes);
-                        return CoreUtil.ToUnixSeconds(result);
-                    },
-                    minArgs: 1,
-                    maxArgs: 2,
-                    isVolatile: true // can depend on current time when called with a single argument
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "AddHours",
-                    category: "NINA",
-                    description: "Adds hours to a datetime. Returns a Unix timestamp in seconds.",
-                    usageExample: "AddHours(now(), 1.5)",
-                    implementation: args => {
-                        DateTime baseDt;
-                        double hours;
-
-                        if (args.Parameters.Length == 1) {
-                            // base: now, arg: hours
-                            baseDt = DateTime.Now;
-                            hours = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        } else {
-                            // first arg: unix timestamp, second: hours
-                            var baseSeconds = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                            baseDt = CoreUtil.UnixTimeStampToDateTime(baseSeconds).ToLocalTime();
-                            hours = Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        }
-
-                        var result = baseDt.AddHours(hours);
-                        return CoreUtil.ToUnixSeconds(result);
-                    },
-                    minArgs: 1,
-                    maxArgs: 2,
-                    isVolatile: true
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "SecondsSince",
-                    category: "NINA",
-                    description: "Returns the number of seconds elapsed since the given datetime (Unix timestamp, seconds).",
-                    usageExample: "SecondsSince(lastEventTime) > 600",
-                    implementation: args => {
-                        var dt = GetDateTime(args); // arg[0] interpreted as unix seconds
-                        var now = DateTime.Now;
-                        return (now - dt).TotalSeconds;
-                    },
-                    minArgs: 1,
-                    maxArgs: 1,
-                    isVolatile: true
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Defined",
-                category: "NINA",
-                description: "Returns whether a symbol name is defined in the symbol table.",
-                usageExample: "Defined(\"foo\")",
-                implementation: args => {
-                    var str = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                    return TryGetValue(str, out _);
-                },
-                minArgs: 1,
-                maxArgs: 1,
-                isVolatile: true // depends on symbol table
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "StartsWith",
-                category: "NINA",
-                description: "Returns whether the string starts with the specified prefix.",
-                usageExample: "StartsWith(\"hello\", \"he\")",
-                implementation: args => {
-                    var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                    var prefix = Convert.ToString(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                    return s.StartsWith(prefix, StringComparison.Ordinal);
-                },
-                minArgs: 2,
-                maxArgs: 2
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "StrLength",
-                category: "NINA",
-                description: "Returns the length of the given string, or -1 if the argument is not a string.",
-                usageExample: "StrLength(\"hello\")",
-                implementation: args => {
-                    var v = args.Parameters[0].Evaluate();
-                    return v is string s ? s.Length : -1;
-                },
-                minArgs: 1,
-                maxArgs: 1
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "StrConcat",
-                category: "NINA",
-                description: "Concatenates two strings.",
-                usageExample: "StrConcat(\"hello\", \" world\")",
-                implementation: args => {
-                    var a = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                    var b = Convert.ToString(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                    return string.Concat(a, b);
-                },
-                minArgs: 2,
-                maxArgs: 2
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "StrAtPos",
-                category: "NINA",
-                description: "Returns the character at the specified zero-based index in a string, or an empty string if the index is out of bounds.",
-                usageExample: "StrAtPos(\"hello\", 1)",
-                implementation: args => {
-                    var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                    var idxObj = args.Parameters[1].Evaluate();
-                    if (idxObj is int idx && idx >= 0 && idx < s.Length)
-                        return s[idx].ToString();
-                    return string.Empty;
-                },
-                minArgs: 2,
-                maxArgs: 2
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Contains",
-                    category: "NINA",
-                    description: "Returns whether the string contains the specified substring.",
-                    usageExample: "Contains(filterName, \"Ha\")",
-                    implementation: args => {
-                        var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        var sub = Convert.ToString(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        return s?.Contains(sub, StringComparison.Ordinal) ?? false;
-                    },
-                    minArgs: 2,
-                    maxArgs: 2
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "EndsWith",
-                    category: "NINA",
-                    description: "Returns whether the string ends with the specified suffix.",
-                    usageExample: "EndsWith(fileName, \".fits\")",
-                    implementation: args => {
-                        var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        var suffix = Convert.ToString(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        return s?.EndsWith(suffix, StringComparison.Ordinal) ?? false;
-                    },
-                    minArgs: 2,
-                    maxArgs: 2
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Substring",
-                    category: "NINA",
-                    description: "Returns a substring of the given string. Out-of-range results in an empty string.",
-                    usageExample: "Substring(\"hello\", 1, 3)  // \"ell\"",
-                    implementation: args => {
-                        var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture) ?? string.Empty;
-                        int start = Convert.ToInt32(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-
-                        if (start < 0 || start > s.Length)
-                            return string.Empty;
-
-                        if (args.Parameters.Length == 2) {
-                            // from start to end
-                            return s.Substring(start);
-                        } else {
-                            int length = Convert.ToInt32(args.Parameters[2].Evaluate(), CultureInfo.InvariantCulture);
-                            if (length < 0)
-                                return string.Empty;
-                            if (start + length > s.Length)
-                                return string.Empty;
-                            return s.Substring(start, length);
-                        }
-                    },
-                    minArgs: 2,
-                    maxArgs: 3
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "ToLower",
-                    category: "NINA",
-                    description: "Converts the string to lower case (invariant).",
-                    usageExample: "ToLower(filterName)",
-                    implementation: args => {
-                        var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        return s?.ToLowerInvariant();
-                    },
-                    minArgs: 1,
-                    maxArgs: 1
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "ToUpper",
-                    category: "NINA",
-                    description: "Converts the string to upper case (invariant).",
-                    usageExample: "toUpper(filterName)",
-                    implementation: args => {
-                        var s = Convert.ToString(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        return s?.ToUpperInvariant();
-                    },
-                    minArgs: 1,
-                    maxArgs: 1
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                name: "Random",
-                category: "NINA",
-                description: "Returns a random double value in the range 0.0–1.0.",
-                usageExample: "Random()",
-                implementation: args => rng.NextDouble(),
-                minArgs: 0,
-                maxArgs: 0,
-                isVolatile: true
-            ));
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Mod",
-                    category: "NINA",
-                    description: "Returns the mathematical modulus (remainder) of x / y. For positive y, the result is in [0, y).",
-                    usageExample: "Mod(-1, 10)",
-                    implementation: args => {
-                        var a = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        var b = Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        var r = a % b;
-                        if (r < 0 && b > 0) r += b;
-                        return r;
-                    },
-                    minArgs: 2,
-                    maxArgs: 2
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Clamp",
-                    category: "NINA",
-                    description: "Clamps a value to the inclusive range [min, max].",
-                    usageExample: "Clamp(exposure, 1, 600)",
-                    implementation: args => {
-                        var v = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        var lo = Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        var hi = Convert.ToDouble(args.Parameters[2].Evaluate(), CultureInfo.InvariantCulture);
-                        return Math.Clamp(v, lo, hi);
-                    },
-                    minArgs: 3,
-                    maxArgs: 3
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Between",
-                    category: "NINA",
-                    description: "Returns whether x is between min and max (inclusive).",
-                    usageExample: "Between(temperature, -10, 40)",
-                    implementation: args => {
-                        var v = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        var lo = Convert.ToDouble(args.Parameters[1].Evaluate(), CultureInfo.InvariantCulture);
-                        var hi = Convert.ToDouble(args.Parameters[2].Evaluate(), CultureInfo.InvariantCulture);
-                        return v >= lo && v <= hi;
-                    },
-                    minArgs: 3,
-                    maxArgs: 3
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Deg",
-                    category: "NINA",
-                    description: "Converts an angle from radians to degrees.",
-                    usageExample: "Deg(PI() / 2)", // assuming user has a PI or similar
-                    implementation: args => {
-                        var rad = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        return AstroUtil.ToDegree(rad);
-                    },
-                    minArgs: 1,
-                    maxArgs: 1
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Rad",
-                    category: "NINA",
-                    description: "Converts an angle from degrees to radians.",
-                    usageExample: "Rad(90)",
-                    implementation: args => {
-                        var deg = Convert.ToDouble(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        return AstroUtil.ToRadians(deg);
-                    },
-                    minArgs: 1,
-                    maxArgs: 1
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Sum",
-                    category: "NINA",
-                    description: "Returns the sum of all arguments.",
-                    usageExample: "Sum(1, 2, 3)",
-                    implementation: args => {
-                        double sum = 0.0;
-                        for (int i = 0; i < args.Parameters.Length; i++) {
-                            sum += Convert.ToDouble(args.Parameters[i].Evaluate(), CultureInfo.InvariantCulture);
-                        }
-                        return sum;
-                    },
-                    minArgs: 1,
-                    maxArgs: int.MaxValue
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Avg",
-                    category: "NINA",
-                    description: "Returns the average (arithmetic mean) of all arguments.",
-                    usageExample: "Avg(1, 2, 3)",
-                    implementation: args => {
-                        double sum = 0.0;
-                        int count = args.Parameters.Length;
-                        for (int i = 0; i < count; i++) {
-                            sum += Convert.ToDouble(args.Parameters[i].Evaluate(), CultureInfo.InvariantCulture);
-                        }
-                        return sum / count;
-                    },
-                    minArgs: 1,
-                    maxArgs: int.MaxValue
-                )
-            );
-
-            RegisterFunction("NINA",
-                new SymbolFunction(
-                    name: "Not",
-                    category: "NINA",
-                    description: "Returns the logical negation of a boolean value.",
-                    usageExample: "Not(cloudy)",
-                    implementation: args => {
-                        bool v = Convert.ToBoolean(args.Parameters[0].Evaluate(), CultureInfo.InvariantCulture);
-                        return !v;
-                    },
-                    minArgs: 1,
-                    maxArgs: 1
-                )
-            );
+            foreach (var fn in new UtilityFunctions(this)) {
+                RegisterFunction(fn.Category, fn);
+            }
         }
 
         void ISymbolBroker.RegisterFunction(ISymbolProvider symbolProvider, SymbolFunction symbolFunction) {
